fix(M): adding alembic script
