+++ backend/alembic/versions/2025_07_14_1053-3ef04912413c_add_daily_session_tracking_to_user_.py
@@ -1,0 +1,36 @@
+"""add daily session tracking to user profile
+
+Revision ID: 3ef04912413c
+Revises: 802e479cfac2
+Create Date: 2025-07-14 10:53:20.998587
+
+"""
+
+from collections.abc import Sequence
+from typing import Union
+
+import sqlalchemy as sa
+
+from alembic import op
+
+# revision identifiers, used by Alembic.
+revision: str = '3ef04912413c'
+down_revision: Union[str, None] = '802e479cfac2'
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    """Upgrade schema."""
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.add_column('userprofile', sa.Column('sessions_created_today', sa.Integer(), nullable=False))
+    op.add_column('userprofile', sa.Column('last_session_date', sa.Date(), nullable=False))
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    """Downgrade schema."""
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_column('userprofile', 'last_session_date')
+    op.drop_column('userprofile', 'sessions_created_today')
+    # ### end Alembic commands ###
